openthermnumber for target water temperature

+++ esphome/components/opentherm/OpenThermListener.h
@@ -6,6 +6,6 @@ namespace esphome::opentherm {
 class OpenThermListener {
  public:
   virtual ~OpenThermListener() = default;
-  virtual void on_response(uint32_t request, uint32_t response);
+  virtual void on_response(unsigned long request, unsigned long response);
 };
 }  // namespace esphome::opentherm
